fix: Separate server and client components to fix build error
